improove solidity highlighter to catch Structs
